Transfered the TalonFX files from a folder to the normal files
