+++ src/main/java/frc/robot/RobotContainer.java
@@ -4,97 +4,147 @@
 
 package frc.robot;
 
-import edu.wpi.first.math.MathUtil;
-import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
+import com.pathplanner.lib.auto.AutoBuilder;
+import edu.wpi.first.math.geometry.Pose2d;
+import edu.wpi.first.math.geometry.Rotation2d;
+import edu.wpi.first.wpilibj.GenericHID;
+import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
-import frc.robot.Constants.OperatorConstants;
-import frc.robot.subsystems.SwerveDriveSubsystem;
-import frc.robot.subsystems.SwerveModuleIO;
-import frc.robot.subsystems.SwerveModuleIOSparkmax;
+import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine;
+import frc.robot.commands.DriveCommands;
+import frc.robot.generated.TunerConstants;
+import frc.robot.subsystems.drive.Drive;
+import frc.robot.subsystems.drive.GyroIO;
+import frc.robot.subsystems.drive.GyroIOPigeon2;
+import frc.robot.subsystems.drive.ModuleIO;
+import frc.robot.subsystems.drive.ModuleIOSim;
+import frc.robot.subsystems.drive.ModuleIOTalonFX;
+import org.littletonrobotics.junction.networktables.LoggedDashboardChooser;
 
+/**
+ * This class is where the bulk of the robot should be declared. Since Command-based is a
+ * "declarative" paradigm, very little robot logic should actually be handled in the {@link Robot}
+ * periodic methods (other than the scheduler calls). Instead, the structure of the robot (including
+ * subsystems, commands, and button mappings) should be declared here.
+ */
 public class RobotContainer {
-  private final SendableChooser<Command> mAutoChooser = new SendableChooser<>();
+  // Subsystems
+  private final Drive drive;
 
-  private final CommandXboxController m_driverController =
-      new CommandXboxController(OperatorConstants.kDriverControllerPort);
-  private final SwerveDriveSubsystem mDriveSubsystem;
+  // Controller
+  private final CommandXboxController controller = new CommandXboxController(0);
+
+  // Dashboard inputs
+  private final LoggedDashboardChooser<Command> autoChooser;
 
   /** The container for the robot. Contains subsystems, OI devices, and commands. */
   public RobotContainer() {
     switch (Constants.currentMode) {
       case REAL:
-        mDriveSubsystem =
-            new SwerveDriveSubsystem(
-                new SwerveModuleIOSparkmax(
-                    Constants.SparkMaxCANID.kFrontLeftDrive,
-                    Constants.SparkMaxCANID.kFrontLeftTurn,
-                    Constants.DriveConstants.kFrontLeftChassisAngularOffset),
-                new SwerveModuleIOSparkmax(
-                    Constants.SparkMaxCANID.kFrontRightDrive,
-                    Constants.SparkMaxCANID.kFrontRightTurn,
-                    Constants.DriveConstants.kFrontRightChassisAngularOffset),
-                new SwerveModuleIOSparkmax(
-                    Constants.SparkMaxCANID.kRearLeftDrive,
-                    Constants.SparkMaxCANID.kRearLefttTurn,
-                    Constants.DriveConstants.kBackLeftChassisAngularOffset),
-                new SwerveModuleIOSparkmax(
-                    Constants.SparkMaxCANID.kRearRightDrive,
-                    Constants.SparkMaxCANID.kRearRightTurn,
-                    Constants.DriveConstants.kBackRightChassisAngularOffset),
-                0);
+        // Real robot, instantiate hardware IO implementations
+        drive =
+            new Drive(
+                new GyroIOPigeon2(),
+                new ModuleIOTalonFX(TunerConstants.FrontLeft),
+                new ModuleIOTalonFX(TunerConstants.FrontRight),
+                new ModuleIOTalonFX(TunerConstants.BackLeft),
+                new ModuleIOTalonFX(TunerConstants.BackRight));
         break;
+
+      case SIM:
+        // Sim robot, instantiate physics sim IO implementations
+        drive =
+            new Drive(
+                new GyroIO() {},
+                new ModuleIOSim(TunerConstants.FrontLeft),
+                new ModuleIOSim(TunerConstants.FrontRight),
+                new ModuleIOSim(TunerConstants.BackLeft),
+                new ModuleIOSim(TunerConstants.BackRight));
+        break;
+
       default:
-        mDriveSubsystem =
-            new SwerveDriveSubsystem(
-                new SwerveModuleIO() {},
-                new SwerveModuleIO() {},
-                new SwerveModuleIO() {},
-                new SwerveModuleIO() {},
-                0);
+        // Replayed robot, disable IO implementations
+        drive =
+            new Drive(
+                new GyroIO() {},
+                new ModuleIO() {},
+                new ModuleIO() {},
+                new ModuleIO() {},
+                new ModuleIO() {});
         break;
     }
-    // Configure the trigger bindings
-    configureBindings();
 
-    mAutoChooser.setDefaultOption("No auto", null);
-    Command instant =
-        Commands.run(() -> mDriveSubsystem.drive(0.5, 0, 0, true), mDriveSubsystem).withTimeout(3);
-    mAutoChooser.addOption("Instant Auto", instant);
+    // Set up auto routines
+    autoChooser = new LoggedDashboardChooser<>("Auto Choices", AutoBuilder.buildAutoChooser());
 
-    SmartDashboard.putData("Auto Selector", mAutoChooser);
+    // Set up SysId routines
+    autoChooser.addOption(
+        "Drive Wheel Radius Characterization", DriveCommands.wheelRadiusCharacterization(drive));
+    autoChooser.addOption(
+        "Drive Simple FF Characterization", DriveCommands.feedforwardCharacterization(drive));
+    autoChooser.addOption(
+        "Drive SysId (Quasistatic Forward)",
+        drive.sysIdQuasistatic(SysIdRoutine.Direction.kForward));
+    autoChooser.addOption(
+        "Drive SysId (Quasistatic Reverse)",
+        drive.sysIdQuasistatic(SysIdRoutine.Direction.kReverse));
+    autoChooser.addOption(
+        "Drive SysId (Dynamic Forward)", drive.sysIdDynamic(SysIdRoutine.Direction.kForward));
+    autoChooser.addOption(
+        "Drive SysId (Dynamic Reverse)", drive.sysIdDynamic(SysIdRoutine.Direction.kReverse));
+
+    // Configure the button bindings
+    configureButtonBindings();
   }
 
-  private void configureBindings() {
-    m_driverController.button(1).whileTrue(mDriveSubsystem.setXCommand()); // a (green) - while true
-    m_driverController
-        .button(2)
-        .onTrue(mDriveSubsystem.toggleFieldRelative()); // b (red) button - on true
-    m_driverController
-        .button(7)
-        .or(m_driverController.button(8))
-        .onTrue(mDriveSubsystem.resetGyroCommand()); // start and back buttons - on true
+  /**
+   * Use this method to define your button->command mappings. Buttons can be created by
+   * instantiating a {@link GenericHID} or one of its subclasses ({@link
+   * edu.wpi.first.wpilibj.Joystick} or {@link XboxController}), and then passing it to a {@link
+   * edu.wpi.first.wpilibj2.command.button.JoystickButton}.
+   */
+  private void configureButtonBindings() {
+    // Default command, normal field-relative drive
+    drive.setDefaultCommand(
+        DriveCommands.joystickDrive(
+            drive,
+            () -> -controller.getLeftY(),
+            () -> -controller.getLeftX(),
+            () -> -controller.getRightX()));
 
-    mDriveSubsystem.setDefaultCommand(
-        Commands.run(
-            () -> {
-              double forwardSpeed = MathUtil.applyDeadband(-m_driverController.getLeftY(), 0.1);
-              double leftSpeed = MathUtil.applyDeadband(-m_driverController.getLeftX(), 0.1);
-              double counterclockwiseSpeed =
-                  MathUtil.applyDeadband(-m_driverController.getRightX(), 0.1);
+    // Lock to 0° when A button is held
+    controller
+        .a()
+        .whileTrue(
+            DriveCommands.joystickDriveAtAngle(
+                drive,
+                () -> -controller.getLeftY(),
+                () -> -controller.getLeftX(),
+                () -> new Rotation2d()));
 
-              if (Math.abs(forwardSpeed) < 0.1) forwardSpeed = 0;
-              if (Math.abs(leftSpeed) < 0.1) leftSpeed = 0;
-              if (Math.abs(counterclockwiseSpeed) < 0.1) counterclockwiseSpeed = 0;
+    // Switch to X pattern when X button is pressed
+    controller.x().onTrue(Commands.runOnce(drive::stopWithX, drive));
 
-              mDriveSubsystem.drive(forwardSpeed, leftSpeed, counterclockwiseSpeed, true);
-            },
-            mDriveSubsystem));
+    // Reset gyro to 0° when B button is pressed
+    controller
+        .b()
+        .onTrue(
+            Commands.runOnce(
+                    () ->
+                        drive.setPose(
+                            new Pose2d(drive.getPose().getTranslation(), new Rotation2d())),
+                    drive)
+                .ignoringDisable(true));
   }
 
+  /**
+   * Use this to pass the autonomous command to the main {@link Robot} class.
+   *
+   * @return the command to run in autonomous
+   */
   public Command getAutonomousCommand() {
-    return Commands.print("No autonomous command configured");
+    return autoChooser.get();
   }
 }
